fix issues with botorch min to max conversion and posterior optim; refactor to add all the utility functions to utils.py and BoTorchRegressor to PyTorchExtrapolator.py
