+++ .gitignore
@@ -47,3 +47,5 @@ __pycache__/
 .cache/
 *pyc
 *egg-info*
+.venv/
+venv/
