+++ AFL/double_agent/__init__.py
@@ -1,7 +1,7 @@
 from .Pipeline import *
 from .AcquisitionFunction import *
 from .Extrapolator import *
-from .BayesianOptimization import *
+from .PyTorchExtrapolator import *
 from .PairMetric import *
 from .Labeler import *
 from .Preprocessor import *
